+++ docker/services/ruby/apps/talk_to_claude/jupyter_with_claude_app.rb
@@ -6,7 +6,7 @@ class JupyterWithClaude < MonadicApp
   icon = "<i class='fa-solid fa-a'></i>"
 
   description = <<~TEXT
-    This is an application that allows Anthropic Claude to create and read/write Jupyter Notebooks.
+    This is an application that allows Anthropic Claude to create and read/write Jupyter Notebooks. The agent can create a new notebook, add cells to an existing notebook, and run the cells in the notebook. The agent can also read the content of the notebook and check the environment in which the notebook is running. <a href="https://yohasebe.github.io/monadic-chat/#/language-models?id=anthropic" target="_blank">Learn more</a>.
   TEXT
 
   prompt_suffix = <<~TEXT